Install Flask-WTF, change of requirements.txt. Moved everythin from blueprints init to user init. The pages work, but not the adding of users to the database when signing up.
