+++ app/blueprints/user/__init__.py
@@ -1,13 +1,16 @@
-from flask import Blueprint, render_template
-from app.controllers import user_controller as uc
-from app.controllers import event_controller as ec
+from flask import Blueprint, redirect, url_for, render_template
+from flask_login import logout_user, login_required
 
-bp_user = Blueprint('bp_user', __name__)
+bp_user = Blueprint("bp_user", __name__)
 
 
-@bp_user.get('/')
-def user():
-    # De-comment to create database posts!
-    # uc.create_user()
-    # ec.create_event()
-    return render_template('user.html')
+@bp_user.get("/signout")
+def signout_get():
+    logout_user()
+    return redirect(url_for("bp_open.index"))
+
+
+@bp_user.get("/schedules")
+@login_required
+def schedules_get():
+    return render_template("schedules.html")
