+++ app/templates/signin.html
@@ -13,6 +13,6 @@
     <form action="{{ url_for("bp_open.signin_post") }}" method="POST">
         <input type="email" placeholder="Email Address" name="email"/> <br/>
         <input type="password" placeholder="Password" name="password"/> <br/>
-        <input type="submit" value="OK" />
+        <input type="submit" value="Sign in!" />
     </form>
 {% endblock %}
