+++ requirements.txt
@@ -2,9 +2,11 @@ click==8.0.3
 colorama==0.4.4
 Flask==2.0.2
 Flask-Login==0.5.0
+Flask-WTF==1.0.0
 itsdangerous==2.0.1
 Jinja2==3.0.3
 MarkupSafe==2.0.1
 pymongo==4.0.1
 python-dotenv==0.19.2
 Werkzeug==2.0.2
+WTForms==3.0.1
